feat: add global handler + clean up

+++ src/main/java/com/example/pro/repository/AppointmentRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface AppointmentRepository extends CrudRepository<Appointment, Integer> {
+public interface AppointmentRepository extends CrudRepository<Appointment, Long> {
     List<Appointment> findAll();
 
     List<Appointment> findByPractitionerId(int practitionerId);
